+++ include/serializer/byte_array_buffer.hpp
@@ -63,7 +63,7 @@ class ByteArrayBuffer {
   std::size_t Tell() const { return pos_; }
 
   std::size_t size() const { return data_.size(); }
-  int64_t bytes_left() const { return data_.size() - pos_; }  
+  int64_t bytes_left() const { return int64_t(data_.size()) - int64_t(pos_); }  
   byte_array::ByteArray const &data() const { return data_; }
 
  private:
