+++ include/serializer/typed_byte_array_buffer.hpp
@@ -81,7 +81,7 @@ class TypedByte_ArrayBuffer {
   std::size_t Tell() const { return pos_; }
 
   std::size_t size() const { return data_.size(); }
-  int64_t bytes_left() const { return data_.size() - pos_; }    
+  int64_t bytes_left() const { return int64_t(data_.size()) - int64_t(pos_); }    
   byte_array_type const &data() const { return data_; }
 
  private:
